fix: Resolve all clippy warnings and add DB migrations to CI
- Fix manual strip_prefix in auth/mod.rs
- Fix redundant pattern matching in websocket handler
- Allow too_many_arguments on handle_send_message (refactor deferred)
- Apply clippy auto-fixes (redundant closures, needless borrows, range contains)
- Add database migration step to all CI jobs (required by sqlx::query! macros)
- Re-format after clippy fixes

+++ src/server/src/handlers/pages.rs
@@ -993,8 +993,8 @@ pub async fn community_posts(
 
     let posts = sqlx::query(&query)
         .bind(uuid)
-        .bind(limit as i64)
-        .bind(offset as i64)
+        .bind(limit)
+        .bind(offset)
         .fetch_all(&state.db.pool)
         .await
         .unwrap_or_default();
